xrp-kernel: allocate shadow buffer with xvp_allocate
Buffers should be allocated from the shared memory, not from the kernel
heap. Shared memory is guaranteed to be accessible to DSP, it's
per-device and is dedicated to the device buffers, and it's potentially
larger than the kernel heap. Shadow buffer doesn't need to be mapped to
the CPU virtual address space except during copy-in and copy-out.

Signen-off-by: Max Filippov <jcmvbkbc@gmail.com>