xrp-dsp: make library code independent of the configuration
Currently xrp_dsp.o has communication area address pointer in it. It is
defined as a magic constant for the linux driver to fix up or the actual
address for fast simulation use case. That makes it impossible to have a
single reusable DSP XRP library. Get rid of it: make that address
undefined in the library code and define it at the final linking of the
firmware with linker directive --defsym.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>