xrp-kernel: implement address translation CPU->DSP
In order to support non-identity mapped system memory and communication
area there must be a way to translate host physical addresses to DSP
physical addresses. Add a structure and functions to support that
translation. Use "ranges" property of the xrp node for the list of
contiguous translation windows. A nested node inside the xrp node is
needed to do the translation. If either ranges property or nested node
or both are missing then identity mapping is used.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>