xrp-example: export system memory from fast simulation model
Fast simulation DSP side may be used not only with the host-based
client, but also with a client running in different simulator, e.g. with
android running in qemu-system-aarch64.
Sharing system memory doesn't make much sense for the host-based client,
because it allocates user memory from the host. OTOH client running in
the full system simulator allocates user buffers from the simulated
physical memory, which can be shared with the XTSC model.

Export system memory from the XTSC in the same way as shared memory.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>