xrp-dsp: xrp-kernel: allow synchronization w/o reset
DSP side of the fast simulation mode may run continuously while the host
side may come and go. The host side will try to synchronize with the DSP
every time it starts. There's no reason to not allow it. It makes
host-DSP interaction more robust in cases when the host does not control
the DSP hardware.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>