xrp-kernel: make sure buffer group data is reachable
Currently buffer group data that doesn't fit into xrp_dsp_cmd is
collected in kernel buffer and that buffer is passed directly to DSP.
However not all of the physical memory may be visible to DSP, so there
may be no translation for arbitrary kernel buffer, which results in
XRP_NO_TRANSLATION address being passed to DSP, causing DSP firmware to
crash.
Use existing xrp_mapping/alien buffer infrastructure to make shadow copy
of the unmappable kernel buffer and sync it back after the command
completion. Turn XRP_MAPPING_KERNEL into a flag that may be combined
with other mapping types, meaning that source data may be not in the
userspace, but that's ok.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>