xrp-linux-*: fix request thread self-termination
In case the last remaining device/queue reference is held by an event in
the request queue, the queue processing thread releases device/queue and
tries to terminate and join itself. It cannot do that and goes to next
iteration where it dereferences now freed device/queue. Add synchronous
termination mechanism to the request processing thread: in case it
cannot join assume that it's the case of synchronous termination.
Communicate it through the flag pointed to by an additional pointer in
the device/queue structure.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>