xrp-kernel: allow cross-device use of shared memory
XRP devices may use shared memory allocated on different XRP devices,
especially when configured to use CMA and common region of reserved
memory. Recognize and support this case.
Shared memory of an XRP device must be fully accessible to that device,
but doesn't have to be directly accessible to any other XRP device. Use
bounce buffer fallback in this case.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>