xrp-host/thread-xos: make stack size configurable
The default stack size for the XOS thread may be enough for multicore
configurations, when the thread is only used to pump requests through
the hardware queue, but it may not be enough for single configuration,
in which DSP code is run in the context of the request processing
thread.
Make XRP thread stack size configurable at link time through the
xrp_xos_thread_stack_size symbol.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>