xrp-host: reorder status update and command queuing
Status pointer may share cache line with writable buffers or output
data. If it is updated after the command is queued for processing
is is possible that it will be updated after the cache invalidation for
the DSP-writable buffers, but before the DSP puts processed data there.
In that case the shared cache line will be populated with the data that
may not match the result written by the DSP.

Reorder status update and command queuing for processing to avoid it.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>