xrp-host: xrp-kernel: add software-assisted edge IRQ support
XRP modes that involve IRQ delivery into the XTSC model have issue with
edge-triggered interrupts: the delivery is performed by the LUA script
and it is unable to detect the edge in the IRQ regiter reliably.
Introduce new IRQ mode: software-assisted edge (#3) and add support for
it to the standalone host code and simple HW kernel driver.
It is asserted by the host side and is cleared by the LUA script. Host
side busy-waits until it's clear.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>