xrp-kernel: fix allocator thread safety for out-of-kernel builds
Allocator uses mutex and atomic variables to serialize access to the
memory pool and reference-cont allocations, but so far both had trivial
no-op implementations when the code was built out-of-kernel.
Use xrp_thread_impl.h and xrp_atomic.h when building out-of-kernel to
provide working host-specific xrp_mutex and _Atomic definitions.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>