xrp-kernel: support multiple HW queues
Read "queue-priority" property from the device tree and if present pass
this information to the DSP in synchronization stage.
Change IRQ handler so that it checks all HW queues.
Order hardware queue priorities and use software queue priority as an
index in the priority-ordered list of queues. Map all software
priorities that exceed the number of hardware queues to the highest
priority queue.
Introduce IOCTL command flag for priority. Route request into a hardware
queue based on its priority.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>