xrp-example: add multiqueue to standalone DTS
Add queue priority definitions to the xrp nodes.
Use binding that supports IRQ delivery to the DSP in order to be able to
test multiqueue functionality. Move software IRQ area out of the first
DSP comm region to a reserved page at the end of the shared memory.
Update register locations in the DTS and in the LUA script.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>