xrp-example: add test for bounce buffer sharing
Make firmware copy data from a range in the input buffer to a range in
the output buffer and then busy-wait for a bit. Submit buffers that will
need bounce buffers to the first N DSPs giving them distinct input and
output ranges and check that the output buffer is not corrupt.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>