xrp-kernel: add hw_ops::property* interface
Add simple HW-specific configuration interface that mimics
OF/device_property and can be deferred to these interfaces.
The purpose is to enable HW that doesn't have natural place to store
this information and host kernels that are too old and/or don't have
said interfaces.

TODO: make default data-driven implementation that can be used by simple
platform-data-style users.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>